simplify Notes Page
