+++ src/data/db.json
@@ -1,4 +1,13 @@
 {
-  "notes": [],
+  "notes": [
+    {
+      "title": "gfds",
+      "tag": "gads",
+      "text": "gfdsa",
+      "id": "e93bd4bf-1448-4bb1-8eb2-871ba5eab7c5",
+      "created": "2023-05-28T13:14:40.006Z",
+      "edited": "2023-05-28T13:14:40.006Z"
+    }
+  ],
   "classes": []
 }
