--- src/pages/Notes/components/NoteControls/components/NoteFilter/styles.module.scss
@@ -1,5 +1,0 @@
-@media only screen and (max-width: 380px) {
-    .input {
-        width: 100%;
-    }
-}
