+++ src/pages/Notes/index.ts
@@ -1,0 +1,3 @@
+export { default as NotesPage } from "./Notes";
+export { NotePage, EditNotePage } from "./pages/Note";
+export { NewNotePage } from "./pages/NewNote";
