+++ src/pages/Notes/pages/NewNote/index.ts
@@ -1,0 +1,1 @@
+export { default as NewNotePage } from "./NewNote";
