+++ src/pages/Notes/pages/Note/index.ts
@@ -1,0 +1,2 @@
+export { default as NotePage } from "./Note";
+export { EditNotePage } from "./pages/Edit";
