+++ src/pages/Notes/pages/Note/pages/Edit/index.ts
@@ -1,0 +1,1 @@
+export { default as EditNotePage } from "./Edit";
