+++ src/pages/Notes/pages/Note/styles/Note.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../assets/variables";
+@import "../../../../../assets/variables";
 
 .page {
     color: white;
