+++ src/pages/Notes/pages/Note/styles/NoteActions.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../../../assets/variables";
+@import "../../../../../assets/variables";
 
 .btn_container {
     position: relative;
