+++ src/pages/Notes/styles/NoteControls.module.scss
@@ -1,9 +1,9 @@
-@import "../../../../assets/variables";
+@import "../../../assets/variables";
 
 .note_controls {
     min-height: 4em;
     display: flex;
-    align-items: center;
+    align-items: flex-end;
     margin: 0 auto;
     flex-wrap: wrap;
     justify-content: space-between;
@@ -13,6 +13,7 @@
 @media only screen and (max-width: $tablet-breakpoint) {
     .note_controls {
         gap: 2em;
+        align-items: center;
         flex-direction: column;
         justify-content: center;
     }
