+++ src/pages/Notes/styles/Notes.module.scss
@@ -1,4 +1,4 @@
-@import "../../assets/variables";
+@import "../../../assets/variables";
 
 .container {
     max-width: 90vw;
